+++ CHANGELOG.md
@@ -24,6 +24,8 @@ Line wrap the file at 100 chars.                                              Th
 
 ## [Unreleased]
 ### Added
+- Add a switch to turn off system notifications under Preferences in the GUI.
+
 #### Windows
 - Add migration logic to restore lost settings after major Windows update.